fix: rebalance response-stream tests for backend-size limit
Move attachment unit test into run_config_test.go and relocate small renderer
helper tests into run_wire_api_test.go so both files stay within the 1000-line cap.
